+++ lib/blinkbox/zuul/server/models/login_attempt.rb
@@ -1,5 +1,4 @@
 module Blinkbox::Zuul::Server
   class LoginAttempt < ActiveRecord::Base
-    belongs_to :user
   end
 end