+++ lib/blinkbox/zuul/server/models/user.rb
@@ -9,7 +9,6 @@ module Blinkbox::Zuul::Server
     has_many :refresh_tokens
     has_many :clients
     has_many :password_reset_tokens
-    has_many :login_attempts
 
     validates :first_name, length: { within: 1..50 }
     validates :last_name, length: { within: 1..50 }
@@ -42,8 +41,17 @@ module Blinkbox::Zuul::Server
     def self.authenticate(username, password, client_ip)
       return nil if username.nil? || password.nil?
       throttle_login_attempts(username)
+
       user = User.find_by_username(username)
-      successful = user != nil && SCrypt::Password.new(user.password_hash) == password
+      if user
+        successful = SCrypt::Password.new(user.password_hash) == password
+      else
+        # even if the user isn't found we still need to perform an scrypt hash of something to help
+        # prevent timing attacks as this hashing process is the bulk of the request time
+        SCrypt::Password.create("random string")
+        successful = false
+      end
+      
       LoginAttempt.new(username: username, successful: successful, client_ip: client_ip).save!
       successful ? user : nil
     end
@@ -83,14 +91,14 @@ module Blinkbox::Zuul::Server
       end
     end
 
-    def self.throttle_login_attempts(username)
-      recent_attempts = LoginAttempt.where(username: username).limit(5).order(id: :desc)
+    def self.throttle_login_attempts(username, max_attempts: 5, period: 20)
+      recent_attempts = LoginAttempt.where(username: username).limit(max_attempts).order(id: :desc)
       failed_attempts = recent_attempts.take_while { |attempt| !attempt.successful? }
-      if failed_attempts.count == 5
-        period = Time.now - recent_attempts.last.created_at
-        if period < 15
+      if failed_attempts.count == max_attempts
+        failed_period = Time.now - recent_attempts.last.created_at
+        if failed_period < period
           error = TooManyAttempts.new("Too many incorrect password attempts; try again later")
-          error.retry_after = (15 - period).to_i.to_s
+          error.retry_after = period - failed_period
           raise error
         end
       end